Add mock trade helpers and refactor panel
