+++ TASKS.md
@@ -389,7 +389,7 @@ Mark each task as complete only when ALL criteria are met:
 - [x] **Task 7:** Real-Time Alert System (Completed: Browser notifications with sound alerts)
 - [x] **Task 8:** Performance Tracking (Completed: Win rate, P&L, and trade metrics)
 - [x] MarketRegimeDetector unit tests implemented
-- [x] **Task 9:** Quick Action Panel (Completed: Buy/Sell, alerts, position sizing, export)
+- [x] **Task 9:** Quick Action Panel (Completed: Quick-trades helper, Buy/Sell, alerts, position sizing, export)
 - [x] **Task 10:** Data Freshness & Reliability
